requirements.txt
backports.functools-lru-cache==1.5
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2018.11.29
chardet==3.0.4
Click==7.0
colorize==1.0.4
cssselect==0.9.1
cycler==0.10.0
dlib==19.16.0
face-recognition==1.2.3
face-recognition-models==0.3.0
feedfinder2==0.0.1
feedparser==5.1.3
geojson==2.3.0
html5lib==1.0.1
idna==2.8
jellyfish==0.6.1
jieba==0.35
kiwisolver==1.0.1
lxml==3.3.5
matplotlib==2.2.3
mechanize==0.4.0
newspaper==0.1.0.7
numpy==1.16.0
opencv-python==4.0.0.21
Pillow==5.4.1
PTable==0.9.2
pycountry==18.12.8
pyparsing==2.3.1
python-dateutil==2.7.5
pytz==2018.9
PyYAML==3.11
requests==2.21.0
scikit-learn==0.20.2
scipy==1.2.0
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
soupsieve==1.7.3
subprocess32==3.5.3
tldextract==1.5.1
urllib3==1.24.1
utils==0.9.0
webencodings==0.5.1
